requirements.txt
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11.15
django-bootstrap4==0.0.7
django-heroku==0.3.1
django-registration==2.4.1
djangorestframework==3.9.0
gunicorn==19.9.0
heroku==0.1.4
idna==2.8
Pillow==5.3.0
psycopg2==2.7.6.1
pycodestyle==2.4.0
python-dateutil==1.5
python-decouple==3.1
pytz==2018.7
requests==2.21.0
urllib3==1.24.1
whitenoise==4.1.2
